fix: Prevent `EISDIR` error when trying to emit REST API input warnings
`GITHUB_ACTION_PATH` points to a directory, not a filename, causing
`readFile` to throw a `EISDIR, read` error.

This resolves that issue by appending `action.yml` to the `actionPath`.

For a more robust approach, the warning function could check if `actionPath`
is a directory first and, if so, append `action.yml`. But, given `GITHUB_ACTION_PATH`
is always a path as provided by Github, this seems reasonable.